refactor: moved middle to refresh user session to supabase/utils directory
Relates #166
