+++ supabase/utils/middleware.ts
@@ -2,7 +2,6 @@ import { createServerClient, type CookieOptions } from '@supabase/ssr';
 import { type NextRequest, NextResponse } from 'next/server';
 
 export const createClient = (request: NextRequest) => {
-  // Create an unmodified response
   let response = NextResponse.next({
     request: {
       headers: request.headers,
